Fixed static typechecking issues, code reworked to use size_t instead of int for array size.

+++ src/noise_est.h
@@ -10,7 +10,8 @@
 
 #include "common.h"
 
-typedef double (*noise_est_func_t)(const double *ns_ps, int fft_size, double *noise_ps, double SNRseg, int samplerate);
+typedef double (*noise_est_func_t)(const double *ns_ps, size_t fft_size, double *noise_ps, double SNRseg,
+                                   int samplerate);
 
 extern noise_est_func_t parse_noise_est_type(const char *name, bool verbose);
 
@@ -18,14 +19,15 @@ extern char *get_noise_est_name(const char *name);
 
 /* Noise estimation algorithms */
 
-extern double hirsch_estimation(const double *ns_ps, int fft_size, double *noise_ps, double SNRseg, int samplerate);
+extern double hirsch_estimation(const double *ns_ps, size_t fft_size, double *noise_ps, double SNRseg, int samplerate);
 
-extern double vad_estimation(const double *ns_ps, int fft_size, double *noise_ps, double SNRseg, int samplerate);
+extern double vad_estimation(const double *ns_ps, size_t fft_size, double *noise_ps, double SNRseg, int samplerate);
 
-extern double doblinger_estimation(const double *ns_ps, int fft_size, double *noise_ps, double SNRseg, int samplerate);
+extern double doblinger_estimation(const double *ns_ps, size_t fft_size, double *noise_ps, double SNRseg,
+                                   int samplerate);
 
-extern double mcra_estimation(const double *ns_ps, int fft_size, double *noise_ps, double SNRseg, int samplerate);
+extern double mcra_estimation(const double *ns_ps, size_t fft_size, double *noise_ps, double SNRseg, int samplerate);
 
-extern double mcra2_estimation(const double *ns_ps, int fft_size, double *noise_ps, double SNRseg, int samplerate);
+extern double mcra2_estimation(const double *ns_ps, size_t fft_size, double *noise_ps, double SNRseg, int samplerate);
 
 #endif
